Move precompile to packaging pipeline
